PSM-1499 Prevent overrideBaseURL from being set if `existingSecret` setup is used
If `existingSecret` is set, we do not wish to set the `redisBaseURL`
statically as we need to inject the password into it, which we do not
have in static context. Therefore, we only set the this URL when we have
Redis in-cluster as the password is set. Also changed the cases for when
Redis is external to be more explicit in better support for the
`existingSecret` use-case.